requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-ckeditor==6.6.1
django-js-asset==2.1.0
django-tinymce==3.6.1
idna==3.4
Pillow==9.5.0
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.0
urllib3==2.0.3
